requirements.txt

fastapi
uvicorn
python-multipart
pytesseract
pillow
pdf2image
spacy
mysql-connector-python
reportlab
python-docx
cryptography
